requirements.txt
fastapi
hypercorn
PyPDF2
pandas
python-docx
chromadb
ollama
sentence-transformers
requests
rdflib 
transformers
rank_bm25
nltk